+++ web-registry/src/pages/Marketplace/Storefront/hooks/useBuySellOrderSubmit.tsx
@@ -184,7 +184,7 @@ const useBuySellOrderSubmit = ({
                     verticalAlign: 'bottom',
                   }}
                 >
-                  <DenomIcon denom={baseDenom} sx={{ display: 'flex' }} />
+                  <DenomIcon baseDenom={baseDenom} sx={{ display: 'flex' }} />
                 </Box>
               ),
             },