+++ web-registry/src/pages/Projects/hooks/useProjectsSellOrders.types.ts
@@ -2,4 +2,5 @@ export type GECKO_PRICES = {
   regenPrice?: number;
   eeurPrice?: number;
   usdcPrice?: number;
+  evmosPrice?: number;
 };